Made an Image class to process the image.
For now, it should be able to detect the white line for Core, but we need to do some further testing to ensure it is 100% accurate.
It has a method to get the location of the white line, and a method to get the error which is how far from the middle of the image the white line is. By default, an error of 0 will be returned if there is no white line. The method which gets the location of the white line could be improved by introducing a third method which just counts the white pixels in the image.
After making that improvement, the next steps are to extend the class so that it can detect other lines in order to process images from the Completion.
Also added robot.hpp.