Extended Image Processing class
The Image Processing class can now detect multiple lines in the image. In addition to the line ahead which it could already detect, it can now also detect lines at the left and right edges of the image. The methods available to the main program have changed, and are now:
getLineAhead()
getLineLeft()
getLineRight()
These will return a struct detailing if a line was found at all, and the location and error of this line:
.found (true/false)
.location (int)
.error (int)
eg getLineLeft().error will return the error of the line the robot is trying to follow.
The class also now contains a "sensor" struct which can be used to define any row/column of pixels in the image for line detection, in case we want to find lines anywhere else.