Added code which can finish Completion course
Updated code from experiments, team meetings, and lab. The Image Processing class now has a new sensor at the bottom of the camera view. The Motor Control class no longer has the turnLeft and turnRight methods, and instead takes an argument to adjust the direction of the robot, as well as one to adjust the speed of the robot. robot.cpp has the following order of logic:
1. Take a left turn if there is one.
2. Take a right turn if there is one and no line ahead.
3. Follow the line ahead while it's still in view (at the bottom of the camera).
4. Failing the above, turn until a line is found again.
This can successfully finish the Core and Completion courses.